requirements.txt
openai
matplotlib
seaborn
pandas
numpy
scikit-learn
python-dotenv
langchain
langchain-openai
langgraph
pydantic

